requirements.txt
PySide6>=6.6.0
xlwings>=0.30.0
google-genai>=1.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
markitdown>=0.0.1
